data.pde graphs data collected by RocketNXP.pde
The changes made to this file allow you to scroll through different flight logs (.csv) files with the click of a mouse. The axis are scaled automatically to help make it easier to view your data. Short altitude gained? No problem, the scale of the graph changes to show your max altitude as the maximum for the graph, and the axis labels are changed on the fly. Scroll through the samples to see what I mean.